% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RPTT.R
\name{getListPageUrls}
\alias{getListPageUrls}
\title{title: getListPageUrls via boardName}
\usage{
getListPageUrls(boardName)
}
\arguments{
\item{boardName}{blablabla}
}
\description{
des: getListPageUrls via boardName
}
\examples{
listPageUrls = getListPageUrls("Gossiping")[1:5]
View(listPageUrls)
}

