% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RPTT.R
\name{getPostUrls}
\alias{getPostUrls}
\title{title: getPostUrls via listPageUrl}
\usage{
getPostUrls(listPageUrl)
}
\arguments{
\item{listPageUrl}{blablabla}
}
\description{
des: getPostUrls via listPageUrl
}
\examples{
listPageUrls = getListPageUrls("Gossiping")[1:5]
postUrls = unlist(lapply(listPageUrls,getPostUrls))
View(postUrls)
}

