% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RPTT.R
\name{getListPageUrls}
\alias{getListPageUrls}
\title{getListPageUrls}
\usage{
getListPageUrls(boardName)
}
\arguments{
\item{boardName}{The name of a board.}
}
\description{
Get urls of the list pages, each contains certain posts of the input board.
}
\examples{
listPageUrls = getListPageUrls("Gossiping")[1:5]
listPageUrls
}

