% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RPTT.R
\name{getPostData}
\alias{getPostData}
\title{getPostData}
\usage{
getPostData(postUrl)
}
\arguments{
\item{postUrl}{The url of the post.}
}
\description{
Get the content of the post.
}
\examples{
getPostData("https://www.ptt.cc/bbs/Gossiping/M.1431338763.A.1BF.html")
}

