% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RPTT.R
\name{getPostUrls}
\alias{getPostUrls}
\title{getPostUrls}
\usage{
getPostUrls(listPageUrl)
}
\arguments{
\item{listPageUrl}{The url of the list page.}
}
\description{
Get urls of posts that listed in the list page.
}
\examples{
listPageUrls = getListPageUrls("Gossiping")[1:5]
postUrls = unlist(lapply(listPageUrls,getPostUrls))
postUrls
}

